separate out reporting
