+++ commands/src/apply/init.rs
@@ -1,12 +1,12 @@
 use crate::apply::lockfile::ApplyLock;
+use crate::apply::report;
 use camino::Utf8PathBuf;
 use common::constants::APPLY_LOCKFILE;
-use common::types::{ApplyOpts, ApplySummary};
+use common::types::ApplyOpts;
 use doers::types::Applicator;
 use embed::compiler;
 use std::process::ExitCode;
-use std::time::{Duration, Instant};
-use util::metrics;
+use std::time::Instant;
 
 pub fn run(host_file: Option<&Utf8PathBuf>, opts: &ApplyOpts) -> ExitCode {
     let lock = ApplyLock::from(APPLY_LOCKFILE);
@@ -43,7 +43,7 @@ pub fn run(host_file: Option<&Utf8PathBuf>, opts: &ApplyOpts) -> ExitCode {
     match applicator.run(opts) {
         Ok(apply_summary) => {
             let elapsed_time = start_time.elapsed();
-            report_success(&apply_summary, elapsed_time, opts.metrics_to.as_deref());
+            report::success(&apply_summary, elapsed_time, opts.metrics_to.as_deref());
         }
         Err(e) => {
             if let Some(host_file) = host_file {
@@ -53,7 +53,7 @@ pub fn run(host_file: Option<&Utf8PathBuf>, opts: &ApplyOpts) -> ExitCode {
             }
 
             let elapsed_time = start_time.elapsed();
-            report_failure(elapsed_time, opts.metrics_to.as_deref());
+            report::failure(elapsed_time, opts.metrics_to.as_deref());
             exit = ExitCode::FAILURE;
         }
     }
@@ -63,29 +63,4 @@ pub fn run(host_file: Option<&Utf8PathBuf>, opts: &ApplyOpts) -> ExitCode {
     }
 
     exit
-}
-
-fn report_success(summary_total: &ApplySummary, elapsed_time: Duration, metrics_to: Option<&str>) {
-    tracing::info!("Run time: {:.3?}", elapsed_time);
-    tracing::info!(
-        "resources: {}  changes: {}",
-        summary_total.resources,
-        summary_total.changes,
-    );
-
-    if let Some(metrics_host) = metrics_to {
-        match metrics::send_as_influx(Some(summary_total), elapsed_time, metrics_host) {
-            Ok(_) => (),
-            Err(e) => tracing::error!("error sending metrics: {}", e),
-        }
-    }
-}
-
-fn report_failure(elapsed_time: Duration, metrics_to: Option<&str>) {
-    if let Some(metrics_host) = metrics_to {
-        match metrics::send_as_influx(None, elapsed_time, metrics_host) {
-            Ok(_) => (),
-            Err(e) => tracing::error!("error sending metrics: {}", e),
-        }
-    }
 }
