+++ commands/src/apply/mod.rs
@@ -1,2 +1,3 @@
 pub mod init;
 pub mod lockfile;
+pub mod report;
